requirements.txt
asgiref==3.3.1
autopep8==1.4.3
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
contextlib2==0.5.5
decorator==4.4.2
Django==3.1.4
django-cors-middleware==1.3.1
django-debug-toolbar==3.2
django-extensions==1.7.1
django-silk==4.1.0
djangorestframework==3.4.4
factory-boy==2.9.2
Faker==5.0.1
freezegun==0.3.11
gprof2dot==2019.11.30
idna==2.8
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
Jinja2==2.10
line-profiler==3.1.0
MarkupSafe==1.1.1
memory-profiler==0.58.0
mock==2.0.0
networkx==1.11
Nuitka==0.6.10.1
parso==0.7.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.4.1
pkg-resources==0.0.0
prompt-toolkit==3.0.8
psutil==5.7.3
ptyprocess==0.6.0
py-spy==0.3.3
pycodestyle==2.6.0
pydotplus==2.0.2
Pygments==2.0.2
pyinstrument==3.2.0
pyinstrument-cext==0.2.3
PyJWT==1.4.2
pyparsing==2.4.7
python-dateutil==2.4.2
pytz==2020.4
requests==2.21.0
simplejson==3.13.2
six==1.11.0
snakeviz==2.1.0
soupsieve==2.1
sqlparse==0.2.4
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.24.3
wcwidth==0.2.5
